Now DISTINCT keyword is supported!
I made changes in the two files database.py and table.py so as the DISTINCT keyword to be supported.

The distinct keyword is now supported in the following way:

When the user inserts a select query, ex. SELECT DISTINCT dept_name FROM department WHERE ...;
the DISTINCT keyword affects the first column after the DISTINCT keyword. So, the first column after the DISTINCT keyword will have only unique values in the result.

The DISTINCT keyword is not yet supported with the "*"  (all the columns).

There are many comments inside the two modified files that explain the DISTINCT operation.