Distinct feature is now totally finished
Now, the user can ask for distinct values in more than one columns. The distinct keyword is applied in every column that is present in the query (before FROM part). In this way, in the result, there will be no duplicates in the combination of the columns and thus every row in the result will be unique!

Co-Authored-By: Christos-N <61213180+Christos-N@users.noreply.github.com>
Co-Authored-By: PanaMour <60044131+PanaMour@users.noreply.github.com>