Removes quotes from multi-word values
If the value of a condition has whitespaces, the user must use double quotation marks.
If they do, the marks get removed and the value is kept without them.
If they don't, an exception is raised letting them know they have to.
For single-word values, double quotation marks are kept as part of the value. (can be changed)

BTW, the error message could be changed, not good with deciding what to log for it.

Problematic test case (TODO):

- Value has double quotes somewhere inside.

  Eg. "com"p. sci." turns to com"p. sci.