IMPORTANT changes for database.py
When a table is dropped, it is removed from the database's __dir__ as it is no longer an attribute.
Also added the option of saving if the table_name specified is a table.
Lastly, removed the save_as option from the joins function as it is never used (only the save_as of a select query is used).
HOWEVER it was not removed completely, in case any problem arises or if it is advised to be changed back. If not, it should be removed in the future.