NOT and BETWEEN Operators Added
The table.py and misc.py files were mainly changed, so that they recognise both operators and change the condition accordingly.
NOT Operator: The main idea is that we take the condition, check whether the 'not' keyword exists, if yes we remove it, and set a parameter named 'notcheck' to true, so that in the misc.py file the operator is reversed to its negated operator accordingly. Finally we reconstruct the condition and send it back to the table.py file.
BETWEEN Operator: We check two keywords, 'between' and 'and' in the condition, and we split it 2 times in each keyword. We get the 2 values of our wanted range, and check whether each entry in the table is within that range. We return  the correct records to the 'values' list and then add them to the rows.