Implementation of BETWEEN Operator
-In database.py file: we check if condition contains the keyword "between".
If so the condition has the format of "column between value1 and value2"
and thus we assign the column name to the condition_column value
-In table.py file: if condition contains the keyword "between", condition is splited into a list using whitespace as seperator.
If list lenght is different than 5 (correct format is:column between value1 and value2) or 'between' and 'and' aren't in the correct positions,exception is raised.
Else, column values are grouped into a list, and then we define a new list(named "rows") in which each value is greater than or to equal value1
and less than or equal to value2 (value1 and value2 are provided by the user in between statement).
In case of error,we raise ValueError because value1 and value2 are not valid.