Fixed Between-Not Operators Issues, added AND/OR Functionality
Fixed some issues in the between and not operators, added 'and' and 'or' functionality.
AND:  The conditions are put in a list and split accordingly to create an intersection of the rows that correspond with the conditions.
OR: Similar functionality, but the rows are created by checking each index in the list of rows that again correspond with at least one condition (meaning at least one is true) and adding those to the final rows.