Added UNIQUE constraint over non-pk column and Btree Indexing functionality.
UNIQUE constraint: Changed the dictionary of the query plan to specify the column in which the index will be created, then changed the Table object to have the unique attribute.
Btree Indexing functionality: 2 columns were added to the meta-indexes table for the specific column and index type created.
The table.py, database.py and mdb.py files were mainly changed.
*Issue with the B+Tree nodes was found.