Added support for Hash index over the PK or unique column.
*not fully commented yet, needs code clean up
The Hash-Index functionality is based on a HashTable object, created in the hash.py file which uses extendible hashing using the MSB variant and the splitting buckets method.
Also added the ability to create more than one indexes on a table (one column with btree index and one with hash index for example) and the correct index is fetched from the meta_indexes table and used to check for the appropriate select method (select_with_btree or select_with_hash). Otherwise only one index type at a time was supported.